add temporal resolution in gridded
